+++ src/neptune/core/operation_processors/async_operation_processor.py
@@ -38,13 +38,13 @@ from neptune.core.components.abstract import WithResources
 from neptune.core.components.metadata_file import MetadataFile
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.core.components.queue.disk_queue import DiskQueue
+from neptune.core.operation_processors.operation_processor import OperationProcessor
 from neptune.envs import NEPTUNE_SYNC_AFTER_STOP_TIMEOUT
 from neptune.exceptions import NeptuneSynchronizationAlreadyStoppedException
 from neptune.internal.exceptions import NeptuneException
 from neptune.internal.init.parameters import DEFAULT_STOP_TIMEOUT
 from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.operation_logger import ProcessorStopLogger
-from neptune.internal.operation_processors.operation_processor import OperationProcessor
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_full_path,
