+++ src/neptune/core/operation_processors/offline_operation_processor.py
@@ -31,8 +31,8 @@ from neptune.core.components.abstract import WithResources
 from neptune.core.components.metadata_file import MetadataFile
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.core.components.queue.disk_queue import DiskQueue
+from neptune.core.operation_processors.operation_processor import OperationProcessor
 from neptune.internal.operation import Operation
-from neptune.internal.operation_processors.operation_processor import OperationProcessor
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_full_path,
