+++ src/neptune/internal/operation_processors/sync_operation_processor.py
@@ -26,7 +26,7 @@ from neptune.constants import SYNC_DIRECTORY
 from neptune.core.components.abstract import WithResources
 from neptune.core.components.metadata_file import MetadataFile
 from neptune.core.components.operation_storage import OperationStorage
-from neptune.internal.operation_processors.operation_processor import OperationProcessor
+from neptune.core.operation_processors.operation_processor import OperationProcessor
 from neptune.internal.operation_processors.utils import (
     common_metadata,
     get_container_full_path,
